Fixed CRC bug, added --raw (headerless) option, gave all options extended alternative names (e.g. --help), got rid of a dependancy. Now optimising for size by default. Release ver. 1.4.0

+++ trunk/Assembler.h
@@ -120,7 +120,7 @@ private:
 	bool writeMmap;
     bool writeHeader;
     // In-file modifiers
-    int start;
+    u16 start;
     double version;
 };
